requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
gunicorn==20.1.0
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
schedule==1.2.0
sentence-transformers==2.2.2
pandas==2.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.0 